requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-text-splitters
langgraph
langchain-groq
langchain-huggingface
sentence-transformers
langchain-chroma
pinecone
langchain-pinecone
pypdf
pydantic
pydantic-settings
python-dotenv
tiktoken
python-multipart
requests
langgraph-checkpoint-sqlite
aiosqlite
mcp
